moved all scheduler-component inside scheduler folder
