+++ src/frontend/src/app/app.routing.ts
@@ -4,11 +4,11 @@ import { HomeComponent } from './home/index';
 import { LoginComponent } from './login/index';
 import { DashboardComponent } from './dashboard/index';
 import { ConfigurationManagerComponent } from './configuration-manager/index';
-import { ScheduledProcessComponent } from './scheduled-process/index';
+import { ScheduledProcessComponent } from './scheduler/scheduled-process/index';
 import { AssetsComponent } from './assets/index';
 import { AuditLogComponent } from './audit-log/index';
 import { SettingsComponent } from './settings/index';
-import { ServicesHealthComponent } from './services-health/index'
+import { ServicesHealthComponent } from './services-health/index';
 
 const appRoutes: Routes = [
     { path: 'home', component: HomeComponent},
