+++ src/frontend/src/app/scheduler/scheduled-process/scheduled-process.component.css
